# specs/003-ai-chatbot-todo/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo AI Chatbot API
  description: |
    AI-powered conversational todo application API with natural language task management.

    Phase III adds conversational interface using OpenAI Agents SDK and MCP (Model Context Protocol).
    Users can manage tasks through natural language conversations.

    **Key Features:**
    - Natural language task management
    - Conversational AI interface
    - MCP-based tool architecture
    - Stateless server design
    - Conversation persistence
  version: 2.0.0
  contact:
    name: Todo Application Support
    email: support@todoapp.com
servers:
  - url: http://localhost:8001
    description: Development server
  - url: https://api.todoapp.com
    description: Production server
paths:
...
  /api/{user_id}/tasks:
    get:
      tags:
        - Tasks
      summary: List all tasks for the authenticated user
      description: Returns all tasks for the specified user, ordered by creation date (newest first).
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user whose tasks to retrieve
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          description: Filter tasks by status (all, pending, completed)
          schema:
            type: string
            enum: [all, pending, completed]
            default: all
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - user does not have access to these tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []

    post:
      tags:
        - Tasks
      summary: Create a new task for the authenticated user
      description: Creates a new task for the specified user.
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user creating the task
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - user does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /api/{user_id}/tasks/{task_id}:
    get:
      tags:
        - Tasks
      summary: Get a specific task by ID
      description: Returns a specific task for the specified user.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []

    put:
      tags:
        - Tasks
      summary: Update a task's title and/or description
      description: Updates an existing task for the specified user.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []

    delete:
      tags:
        - Tasks
      summary: Delete a task permanently
      description: Deletes a task for the specified user.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []
  /api/{user_id}/tasks/{task_id}/complete:
    patch:
      tags:
        - Tasks
      summary: Toggle task completion status
      description: Toggles the completion status of a task.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
              required:
                - completed
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []

  # ==========================================
  # PHASE III ENDPOINTS (New)
  # ==========================================
...
components:
  schemas:
...
    TaskCreate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title (required)
          example: Buy groceries
        description:
          type: string
          maxLength: 1000
          description: Task description (optional)
          example: Milk, eggs, bread
      required:
        - title
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Buy groceries and cook dinner
        description:
          type: string
          maxLength: 1000
          example: Milk, eggs, bread, and ingredients for dinner
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          example: Buy groceries
        description:
          type: string
          nullable: true
          example: Milk, eggs, bread
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: 2026-01-19T10:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2026-01-19T10:00:00Z

    # ==========================================
    # CHAT SCHEMAS (Phase III)
    # ==========================================
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Invalid email or password
        code:
          type: string
          description: Machine-readable error code
          example: INVALID_CREDENTIALS
        details:
          type: object
          nullable: true
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          example: 2026-01-26T10:00:00Z

  # ==========================================
  # REUSABLE RESPONSES
  # ==========================================
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - user does not have access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found - resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # ==========================================
  # SECURITY SCHEMES
  # ==========================================
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token required for authentication.

        Obtain token via /api/auth/signup or /api/auth/signin endpoints.

        Include in Authorization header: `Authorization: Bearer <token>`
security:
  - bearerAuth: []
...